+++ src/TestEngine/nunit-agent/Program.cs
@@ -31,6 +31,7 @@ using NUnit.Engine;
 using NUnit.Engine.Agents;
 using NUnit.Engine.Internal;
 using NUnit.Engine.Services;
+using NUnit.Engine.Utilities;
 
 namespace NUnit.Agent
 {
