+++ src/TestEngine/nunit-agent/nunit-agent.csproj
@@ -13,7 +13,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="..\EngineVersion.cs" LinkBase="Properties" />
-    <Compile Include="..\nunit.engine.core\Internal\ExceptionHelper.cs" Link="ExceptionHelper.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="..\..\..\nunit.ico">
