+++ src/TestEngine/nunit.engine.core/Agents/RemoteTestAgent.cs
@@ -26,8 +26,8 @@ using System;
 using System.Runtime.Remoting.Channels;
 using System.Runtime.Remoting.Channels.Tcp;
 using System.Threading;
-using NUnit.Common;
 using NUnit.Engine.Internal;
+using NUnit.Engine.Utilities;
 
 namespace NUnit.Engine.Agents
 {
