+++ src/TestEngine/nunit.engine.core/Internal/DomainDetailsBuilder.cs
@@ -1,4 +1,4 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) 2017 Charlie Poole, Rob Prouse
 //
 // Permission is hereby granted, free of charge, to any person obtaining
@@ -26,9 +26,9 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Text;
-using NUnit.Common;
+using NUnit.Engine.Internal;
 
-namespace NUnit.Engine.Internal
+namespace NUnit.Engine.Utilities
 {
     /// <summary>
     /// DomainDetailsBuilder provides human readable information on
