+++ src/TestEngine/nunit.engine.core/Internal/TcpChannelUtils.cs
@@ -29,7 +29,7 @@ using System.Runtime.Remoting.Channels;
 using System.Runtime.Remoting.Channels.Tcp;
 using System.Runtime.Serialization.Formatters;
 using System.Threading;
-using NUnit.Common;
+using NUnit.Engine.Utilities;
 
 namespace NUnit.Engine.Internal
 {
