+++ src/TestEngine/nunit.engine.core/Utilities/ExceptionHelper.cs
@@ -1,4 +1,4 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) 2017 Charlie Poole, Rob Prouse
 //
 // Permission is hereby granted, free of charge, to any person obtaining
@@ -25,11 +25,10 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Text;
-using NUnit.Engine;
 
-namespace NUnit.Common
+namespace NUnit.Engine.Utilities
 {
-    internal static class ExceptionHelper
+    public static class ExceptionHelper
     {
         /// <summary>
         /// Builds up a message, using the Message field of the specified exception
